+++ tools/mock_node/src/setup.rs
@@ -12,7 +12,7 @@ use near_client::{start_client, start_view_client, ClientActor, ViewClientActor}
 use near_epoch_manager::EpochManager;
 use near_network::types::NetworkClientMessages;
 use near_network::types::NetworkRecipient;
-use near_primitives::state_part::PartId;
+use near_primitives::state::PartId;
 use near_primitives::syncing::get_num_state_parts;
 use near_primitives::types::BlockHeight;
 use near_store::test_utils::create_test_store;
